add output and handling when failing to reach the github files

+++ install.sh
@@ -172,11 +172,10 @@ Press ENTER to continue, or CTRL+C to cancel. "
 clear -x
 
 installation_fail () {
-  docker compose --profile prod logs
-  read -r -p "
+  printf "
 Unfortunately, the installation has failed.
 
-We've printed some error logs above that will hopefully point you to the
+We'll print some error logs above that will hopefully point you to the
 problem.
 
 A common issue is for there to be some networking issue outside of Mathesar's
@@ -185,11 +184,18 @@ control. Please:
 - Make sure you have access to https://raw.githubusercontent.com/
 
 If you can't get things working, please raise an issue at
-https://github.com/centerofci/mathesar/issues/
 
-Press ENTER to reset the local docker environment. "
-  docker compose --profile prod down -v --rmi all
-  read -r -p "Press ENTER to exit the installer. "
+https://github.com/centerofci/mathesar/issues/
+" >&2
+
+  if [ "${1}" == "late" ]; then
+    read -r -p "
+    Press ENTER to print the logs and reset the local docker environment. "
+    docker compose --profile prod logs
+    docker compose --profile prod down -v --rmi all
+  fi
+  read -r -p "
+Press ENTER to exit the installer. "
   exit 1
 }
 
@@ -393,14 +399,14 @@ installation.
 "
 printf "Downloading docker-compose.yml...
 "
-sudo curl -sL -o docker-compose.yml https://raw.githubusercontent.com/centerofci/mathesar/"$github_tag"/docker-compose.yml
+sudo curl -sfL -o docker-compose.yml https://raw.githubusercontent.com/centerofci/mathesar/"${github_tag}"/docker-compose.yml || installation_fail early
 read -r -p "Success!
 
 Next, we'll download files and start the server, This may take a few minutes.
 
 Press ENTER to continue. "
 clear -x
-docker compose --profile prod up -d --wait || installation_fail
+docker compose --profile prod up -d --wait || installation_fail late
 clear -x
 printf "
 --------------------------------------------------------------------------------
